+++ devtools/common/src/commonTest/kotlin/com/maximbircu/devtools/common/core/DevToolTest.kt
@@ -18,7 +18,7 @@ class DevToolTest : BaseTest() {
     }
 
     @Test
-    fun `some test`() {
+    fun `preferences dev tool is using a preferences tool store`() {
         mockkObject(PreferencesToolStore)
         every { PreferencesToolStore.create<Any>(any()) } returns mockk(relaxed = true)