+++ samples/android/src/main/kotlin/com/maximbircu/devtools/MainActivity.kt
@@ -2,6 +2,12 @@ package com.maximbircu.devtools
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.maximbircu.devtools.configscreens.CombinedSourcesConfigActivity
+import com.maximbircu.devtools.configscreens.JsonSchemaSourceConfigActivity
+import com.maximbircu.devtools.configscreens.MemorySourceConfigActivity
+import com.maximbircu.devtools.configscreens.YamlSourceConfigActivity
+import kotlinx.android.synthetic.main.activity_main.combinedSourceTools
+import kotlinx.android.synthetic.main.activity_main.jsonSchemaSourceTools
 import kotlinx.android.synthetic.main.activity_main.memorySourceTools
 import kotlinx.android.synthetic.main.activity_main.yamlSourceTools
 
@@ -10,7 +16,9 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        memorySourceTools.setOnClickListener { MemoryToolsActivity.start(this) }
-        yamlSourceTools.setOnClickListener { YamlToolsActivity.start(this) }
+        memorySourceTools.setOnClickListener { MemorySourceConfigActivity.start(this) }
+        yamlSourceTools.setOnClickListener { YamlSourceConfigActivity.start(this) }
+        jsonSchemaSourceTools.setOnClickListener { JsonSchemaSourceConfigActivity.start(this) }
+        combinedSourceTools.setOnClickListener { CombinedSourcesConfigActivity.start(this) }
     }
 }